Build 10586
-Prevent logging NCI data messages unless a reg key is set
-Fix bug where the driver doesn't properly restore the states of the
secure elements (SE's). If the driver encounters a failure and tries to
recover, it currently disables all the SE's as part of the sequence.
Clients are not notified of this, and therefore still believe the SE's
are in their original state. The fix changes the code so the SE's are
reset to their correct state.