Rework CORE_RESET_NTF Callback (phNciNfc_ResetNtfCb)
This rework fix the way CORE_RESET_NTF is handled.
A CORE_RESET_NTF (in NCI1x or NCI2x) can be raised by the NFC Controller
any time in the NCI state machine (i.e: In case of CLF error, if the NFCC
was powered on (NCI2x), if a CORE_RESET_CMD was received...).
In any case, the overall state machine needs to be resetted (NCI state
machine, IoQueue...).