Fix sequence executed after receiving a DISCOVER_TARGET_ACTIVATION_FAILED error
In NCI1x or NCI2x we can receive
CORE_GENERIC_ERROR_NTF(DISCOVER_TARGET_ACTIVATION_FAILED) in 2 states:
- RFST_DISCOVERY (in this case the current code is fine).
- RFST_W4_HOST_SELECT (in this case the current code does not restart the
polling loop).

Following the 5.2 State machine and the NfcCx commands exchanges when using
a tag (for instance in our tests Mifare Ultralight):

- in RFST_DISCOVERY state: when approching a tag a RF_INTF_ACTIVATED_NTF will
change the current state to RFST_POLL_ACTIVE.

- In RFST_POLL_ACTIVE state: the NfcCx send a RF_DEACTIVATE_CMD(Sleep Mode).
This is changing the state to RFST_W4_HOST_SELECT.

- In RFST_W4_HOST_SELECT: A CORE_GENERIC_ERROR_NTF(DISCOVER_TARGET_ACTIVATION_FAILED)
may happen.
According to 5.2.4:
"If the activation was not successful, the NFCC SHALL send CORE_ERROR_NTF to
the DH with a Status of DISCOVERY_TARGET_ACTIVATION_FAILED and the state will
remain as RFST_W4_HOST_SELECT"
"If the DH sends RF_DEACTIVATE_CMD,
the NFCC SHALL ignore the Deactivation Type parameter, stop the RF Discovery
Process and send RF_DEACTIVATE_RSP. The state will then change to RFST_IDLE."

- In RFST_IDLE: no field is generated until the NCI state machine is change
to RFST_DISCOVERY.

In order to consider both cases we are now sending the NFC controller in
RFST_IDLE state and move to RFST_DISCOVERY with a RF_DISCOVER_CMD.