Add Alternative way of sending a Mifare Write or IncDecRestoreTransfer Command
The existing function allowing to send a Mifare Write16 or a
IncdevRestoreTransfer command are done in a way that the full data is sent
in 2 steps (e.g: command header <cmd><addr> and then payload in 2 differents
nci command).
STMicroelectronics chipset support only sending those in one nci command.
Todo as an improvement remove the former method of sending.