Add support for Mifare Classic Presence check for STM NFC Controller
Mifare Classic Presence Check relies on authentication command.
The command format used so far is proprietary and STM NFC Controller is
using a different format. Mainly the UID is missing in the existing command.
Save the UID in the MFCKey if we use a STM NFC Controller and generate
the Presence check command accordingly to the Mifare authentication command.