Add support for STM NFC Controller Mifare Classic NCI data response
The NCI data response for a Mifare Classic tag is a bit different for NXP.
For STM, Mifare Classic commands are Raw based and the answer a little
different.
If data are present (> 2), no status byte is present.
If data == 2 (in case of failure only)
 format is <NCI FAILURE> <mifare status byte>
if data == 1 (in case of success only) <NCI_SUCCESS)