Allow the NFC radio to turn off if there aren't any pending smartcard IOCTLs.
Currently the NFC radio is kept on whenever there is an open smartcard
handle. This change changes this to only requiring either a pending
IOCTL or a connected smartcard (i.e. a card in the NFC radio field).

The IOCTLs that are of interest are IS_ABSENT, IS_PRESENT and TRANSMIT.
All the other IOCTLs don't require the NFC radio. In addition, TRANSMIT
already has its own power management logic and this doesn't need to be
changed.

IS_ABSENT and IS_PRESENT used to use a WdfIoQueue to both track the
pending IOCTL requests and to automatically handle request cancelation.
However to be able to be handle power references correctly, without race
conditions, I needed to change this to manually track the requests and
cancelation.

As part of this change, a de-bounce was added to powering down the RF.
This was added to prevent powering on and off the RF quickly within a
short timeframe.

In addition, this change simplifies the way the RF Interface receives
the desired state of the RF from the Power Manager. This was done to
remove race conditions and improve performance.

Miscellanious changes:
- Removed the NCI_POWER_POLICY type, as it was just adding unneccessary
boilerplate.