Handle STATUS_PENDING status code when calling 'WdfDeviceStopIdle'.
The 'WdfDeviceStopIdle' function returns STATUS_PENDING when the device
is booting up asynchronously. As STATUS_PENDING is a success code (not
an error code) this generally doesn't cause too many problems. However
STATUS_PENDING does have a special meaning in some circumstances. So to
be safe, we should convert it to STATUS_SUCCESS.