Fix Type 1 tag CC version check (#25)
When attempting to write and NDEF to a Type 1 (Topaz) tag, the stack is
checking the capability container (CC) version (byte 1) has a value of
0x10 (i.e. version 1.0). However, a type 1 tag could also be version 1.1
or 1.2 at this point in time. A version 1.x tag will be compatible with
a stack using 1.0 features. Therefore, the version check should ignore
minor version and only validate the major version is 1.