Fix the way a secure element is detected.
Following NCI2.0 specification NFCEE_DISCOVER_NTF description and figure 25,
A NFCEE_DISCOVER_NTF signal a supported NFCEE but does not signal a physically
connected/present NFCEE.
Figure 25 shows that a NFCEE_TRANSMISSION_ERROR signal an unresponsive NFCEE
or not connected/present.
As a consequence, the NFCEE is removed from the NFCEE list.