Ensure Airplane Mode setting is remembered accross NFC driver restarts.
In a previous change, the logic that handled restoring the previous
Airplane Mode setting was moved into the Power Manager class. However
the registry values that cache the previous Airplane mode are currently
read before the Power Manager class is initialized.

This change ensure the registry values are read before the Power Manager
is initialized.