Ensure global variables are only accessed through getters/setters.
This change also ensures that a global variable's value is only
retrieved a single time within a function and is cached in a local
variable.

This refactoring will make it easier to fix the global variable race
conditions bugs that exist. Ideally we would modify the code so that
there aren't any global variables and have everything passed as function
parameters. However we could also use TLS values to fix these problems.