Add and implement IOCTL_NFCSE_SET_POWER_MODE.
IOCTL_NFCSE_SET_POWER_MODE will be used to request the NFC Controller
to power an SE (e.g. eSE) even if the NFC Controller is not using the
eSE itself. This will primarily be used situations when there is a
secondly communication channel with the SE that is unable power the SE
directly.