Ensure eSE remains powered between multiple transmit commands.
On some NFC Controllers, the eSE is completley powered off after a
command hasn't been sent to it for a period of time. For these NFC
Controllers, we need to ensure that the eSE remains powered while there
is an active client of the eSE. This can be achieved through using the
Power and Link Control command, if it is avaliable.

The change also uses the Power and Link Control in the
IOCTL_NFCSE_SET_POWER_MODE API for similar reasons.