Stop power references from underflowing in HLK test.
1. Ensure that setting an SEs activation mode is not skipped
incorrectly. (This prevents the activation mode from getting out-of-sync
with the emulation mode.)

2. Ensure the SEs emulation mode is always set to off when a
SEManager client handle is closed, even if the activation mode and
emulation mode are out-of-sync. This is done just in case there is
another way an SEs activation mode and emulation mode can get
out of sync.