Add new event types for RF field entry/exit events.
Currently within NfcCx, the RF field entry/exit events only fire once a
connection to the NFC terminal has been established. This is problematic
in the case when 'EmulationStealthListen' is being used, as the sole
purpose of that emulation mode is to prevent a connection from being
established while still permitting the RF field entry/exit events to
fire.

This change adds two new event types to the SE driver API that report
the true field entry/exit events.