Improve HCI static connection credits management on CORE_INIT_RSP
CORE_INIT_RSP may have Initial Credit Value set to 0 or a non 0 value.
Only update the current HCI credit value to the CORE_INIT_RSP Initial
Credit Value and move to the next sequence.
If the Initial Credit Value is 0,
phNciNfc_ProcessConnCreditNtf will update the credit for the HCI static
connection when it is ready.
Also, change the way we are checking if the HCI static connection is
implemented using b3 of the DiscConfSuprt byte.