Fix eSE power.
1. Allow NCI 1.x devices to specify that support the Power and Link
Control command.

2. Ensure that the Power and Link Control command is sent before the
SE Set Mode command.

3. In some circumstances changing the card emulation or changing an SE's
power mode can fail for unexpected reasons. This can result in the power
ref-counts from getting out of sync, which causes the driver to fail.

To fix this issue, change the card emulation power ref-count and the SE
power ref-count to be global power ref-counts instead of ref-counts
associated with a file handle.