Ensure IFD GetAtr does not require a card reset.
This change ensures that a call to retrieve the ATR through
IOCTL_SMARTCARD_GET_ATTRIBUTE doesn't invoke a card reset,
as such behavior is not compliant with the smartcard driver interface
and can cause client scenarios to fail unexpectedly.

Instead the IFD ATR is cached after all successful card
reset operations. This cached value is then used for any
IOCTL_SMARTCARD_GET_ATTRIBUTE ATR requests. In addition, a card reset
operation is run when on the first client connection to the eSE to
ensure the ATR value is available if and when it is requested.