Ensure a slow NCI write I/O request doesn't cause state corruption.
In the current code, when an NCI write I/O request is made and a
response packet is expected, the response watchdog timer is started when
the NCI write is sent instead of when it has completed. In most cases,
this is fine as the NCI write is expected to only take a short amount of
time and therefore is likely to finish well before the response watchdog
timer fires. However on some hardware, it seems that the NCI write can
take a somewhat long time to complete in certain situations. When this
occurs, the driver code treats the error like any other error and
resumes what it was doing. However, since a new NCI write can't be sent
before the old one has completed, any new NCI commands fail which can
cause the various state machines to become confused.

To fix this issue, the response watchdog timer was changed so that it is
only started after the NCI write has completed (instead of when the NCI
write is sent).