Refactor the NfcCxRF state handling.
The primary purpose of this change is to ensure that a state transition
caused by internal driver event (e.g. a user tapping a tag) doesn't
interfere with an already running state transition (if any). There is
already logic that delays user events (e.g. a tag read I/O request)
for the same reason. This change rewrites this logic so that it supports
both internal and user events.

Unfortunatley, the existing logic that delays users events used "magic"
to determine when an event should be delayed and when it should be
handled immediately. This made the state machine logic an absolute
nightmare to read and comprehend. In addition, updating the "magic"
logic to handle internal events and user events would have been almost
impossible.

This change rips out the "magic" logic and instead requires the callers
that are issuing the events to explictly state if the event is
deferrable or not.

Miscellaneous changes:
- Fix `NfcCxRFInterfaceExecute` so that it only needs a single event
handle instead of two.