Rewrite NfcCxRF sequence handling.
1. Remove the concept of sequence process functions. These functions
were originally intended to be used to process the results of
asynchronous operations. However the result processing logic instead
exists in the operations' callback handlers. So the sequence process
functions are always set to NULL and are therefore pointless.

2. Store the value of the sequence complete function (the function
called when the sequence has finished) separately, instead of having it
be part of the the list of sequence handlers. This was done because
there is a number of sequences that only differ by their complete
function. So it makes the code cleaner to support this directly.

3. Ensure that all sequence lists are specified in immutable global
memory, instead of mutable global memory. (#2 made doing this a lot
easier.)

4. Store the state of a running sequence in heap allocated memory. This
memory is used:
    1. To uniquely identify a running sequence, so that sequence
    conflicts can be detected.
    2. As the context value for (LibNfc) asynchronous operations. This
    replaces the mutable global variables that were previously used for
    this purpose. (e.g. NFCCX_RF_LIBNFC_REQUEST_CONTEXT.)

5. Make the sequence complete functions return `void` instead of
`NTSTATUS`. This error code was never used for anything. So its
existence only causes confusion.

6. Remove the `Status` parameter from the sequence handler functions.
This value was alaways `STATUS_SUCCESS` and was never used for anything.
So its existence only causes confusion.

7. If the initial sequence handler completes synchronously (either
success or failure), then call complete function in a new message to the
LibNfc thread. This ensures that the state handler function is not
called recursively.