UT: Make it easier to compose sequences together.
Currently the test suite requires sequence steps to be specified in a
single continuous array. However it would be useful to have some easy
way to concatenate sequences together so that the sequences can be
modular.

While it would be possible to do this concatenation at runtime, this
would come at a runtime cost. So extra code would be required to
mitigate these costs to prevent them from becoming problematic. For
example, in the airplane mode tests the `InitializeNoSEs`,
`DiscoveryStart`, `DiscoveryStop` and `Uninitialize` sequences are all
used multiple times. So some form of caching mechanism would be required
to avoid regenerating the same data over and over again.

To avoid these problems, the sequences will instead be concatenated at
compile time using a tree structure. This has the nice feature of not
duplicating data if one sequence is used by multiple other sequences.