Fix an issue when generating a routing table cmd payload.
As specified by NCI2.0 NFC Forum specification,
"The routing entries sent by the DH in RF_SET_LISTEN_MODE_ROUTING_CMD SHALL
be in the following order:

- All AID-based routing entries (if any)
- All APDU Pattern-based routing entries (if any)
- All System Code-based routing entries (if any)
- All Protocol-based routing entries (if any)
- All Technology-based routing entries (if any)"

The idea of phNciNfc_BuildSetLstnRtngCmdPayload is to generate a payload
following the above statement for AID-based Protocol-based and
Technology-based routing entries.

By browsing s_requiredEntryTypeOrdering, every routing entry shall be loaded
into the command payload in that order.
Only "if (pRtngConfig[entryIndex].Type == currentType)". If not the function
continue browsing.
The