UT: Add ability to wait for device interface to show up.
For some device interfaces (e.g. eSE smartcard), there isn't a signal
from the simulator driver that provides a gaurantee that the device
interface has been created and is ready to be used. So we have to wait
an unspecified amount of time for the interface to show up.

This change modifies the `DeviceQuery::FindDriverInterfaces` function
to take a timeout value. Once at least 1 interface has been found, this
interface is returned immediately. Otherwise the function will wait for
the timout to expire.

As part of this change the `DeviceQuery` was rewritten to use the
`windows.devices.enumeration` APIs as they are more flexible than the
Configuration Manager APIs.