UT: Fix race condition during test initialization.
Currently the test client driver has an implicit assumption that the
first test suite connection will be opened when the driver is not in D0.
If this isn't the case, which can happen if the test suite is started
immediately after the driver is initialized, then the test suite will
hang indefinitely.

To fix this problem, the test driver will no longer start and stop NfcCx
during D0Entry and D0Exit. Instead the test suite will directly control
when NfcCx is started and stopped via IOCTLs. In addition, the test
driver will inform the test suite of the D0Entry and D0Exit events via
callbacks so that the test suite can verify the behavior of the NfcCx
Power Manager.

As part of this change, the `NciSimConnector` callback retrieval was
changed so that it is done asynchronously in the background. This was
done to make it possible to split the LibNfc thread callbacks and the
power manager callbacks, as these callback events are not synchronized
and therefore can arrive in a non-determined order.

In addition, the synchronous commands of `NciSimConnector` were changed
to have a timeout value, to help reduce the likelihood of the test suite
becoming blocked indefinitely.