UT: Install/uninstall test device during test setup/cleanup.
This changes remove the auto installation of the test device when the
test driver is installed. Instead the test suite will install the device
during test initialization and uninstall the test device when the test
uninitializes (including when the test suite crashes).

This is being done to make it easier to re-run the test suite when
debugging test failures. (Previously you would have to manually disable
and enable the test device, to reset the test device's state.) It also
means the system will be left in a cleaner state when the test suite has
finished running, which will be useful in lab test runs which run
multiple test suites on the same VM.