NfcCx: Fix NCI 2 behavior when connecting to a T2T tag.
Currently there is code in the NFC CX that triggers NFC-A's SLP_REQ
command (NFC Forum, Digital Protocol, Version 2.1, Section 6.9.1) when
connecting to a T2T (e.g. Mifare) NFC tag when in NCI 2 mode.

However there are two problems:

1. The NCI specs (both versions 1 and 2) explictly call out that the
SLP_REQ command should be handled by the NFC Controller not the driver.
(NFC Forum, NCI, Version 2.0, Section 8.2.4.2, Table 94.)

2. The wrong command code is being used. Specifically, '05 00' instead
of '50 00'. So even if the SLP_REQ command was required, it would be
broken.

This change removes the code that invokes the SLP_REQ command.

In addition, tests are added for reading a T2T tag in NCI 2 mode. Also,
some code duplication was removed from the `TagTests` class.