Support getting ATR via reset for NFC smartcard.
The IOCTL_SMARTCARD_POWER command (aka, smartcard reset) is supposed to
return the smartcard's ATR in the I/O request's output buffer. However
the NFC CX has a long standing issue where this isn't done for the NFC
smartcard interface. This change fixes this.

In addition, a test is being added to the UT suite that tests the
behavior of the NFC smartcard reset operation, including checking the
ATR is returned. Also, the TagTests class is being refactored slightly
to remove duplicate code.